+++ routes/movies.js
@@ -1,7 +1,7 @@
 import express from 'express';
 
 import { createMovie, getMovies } from '../controllers/movies.js';
-import allMovieFields from './movies.validator.js';
+import allMovieFields from '../validators/movies.js';
 
 const router = express.Router();
