+++ utils/constants.js
@@ -10,19 +10,16 @@ export const HASH_SALT_LENGTH = 10;
 export const JWT_SETTINGS = { expiresIn: '7d' };
 export const JWT_COOKIE_MAX_AGE = 1000 * 60 * 60 * 24 * 7; // 1 week
 
-export const userValidationSettings = {
+export const validationSettings = {
+  STRING_MIN: 2,
+  SHORT_STRING_MAX: 30,
+  LONG_STRING_MAX: 2000,
+  COUNTRY_PATTERN: /[\p{L}\- ()']*/u, // Supports Côte D'Ivoire
   EMAIL_MAX: 128,
   PASSWORD_MIN: 6,
   PASSWORD_MAX: 30,
   NAME_MIN: 2,
   NAME_MAX: 30,
-};
-
-export const movieValidationSettings = {
-  STRING_MIN: 2,
-  SHORT_STRING_MAX: 30,
-  LONG_STRING_MAX: 2000,
-  COUNTRY_PATTERN: /[\p{L}\- ()']*/u, // Supports Côte D'Ivoire
 };
 
 export const statusCodes = {
