+++ validators/user.js
@@ -1,6 +1,6 @@
 import { celebrate, Joi } from 'celebrate';
 
-import { userValidationSettings as settings } from '../utils/constants.js';
+import { validationSettings as settings } from '../utils/constants.js';
 
 const email = Joi.string().email().max(settings.EMAIL_MAX);
 const password = Joi.string().min(settings.PASSWORD_MIN).max(settings.PASSWORD_MAX);
